+++ frontend/components/layout/LocaleMenu.vue
