+++ frontend/layouts/demo.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import TheHeader from '~/components/organisms/layout/TheHeader'
+import TheHeader from '~/components/layout/TheHeader'
 
 export default {
   middleware: ['check-auth'],
