+++ frontend/layouts/project.vue
@@ -37,8 +37,8 @@
 
 <script>
 import { mapGetters } from 'vuex'
-import TheSideBar from '~/components/organisms/layout/TheSideBar'
-import TheHeader from '~/components/organisms/layout/TheHeader'
+import TheSideBar from '~/components/layout/TheSideBar'
+import TheHeader from '~/components/layout/TheHeader'
 
 export default {
   middleware: ['check-auth', 'auth', 'check-admin'],
