+++ frontend/pages/index.vue
@@ -7,9 +7,9 @@
 </template>
 
 <script>
-import FeatureCards from '@/components/organisms/layout/FeatureCards'
-import TheTopBanner from '@/components/organisms/layout/TheTopBanner'
-import TheBottomBanner from '@/components/organisms/layout/TheBottomBanner'
+import FeatureCards from '@/components/layout/FeatureCards'
+import TheTopBanner from '@/components/layout/TheTopBanner'
+import TheBottomBanner from '@/components/layout/TheBottomBanner'
 
 export default {
   components: {
